- build ResultSelfieActivity.kt (to show selfie) - remake SelfieActivity.kt - edit some code on TakeSelfieActivity.kt
