+++ .idea/misc.xml
@@ -6,7 +6,8 @@
         <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/drawable/ic_baseline_camera.xml" value="0.24" />
         <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_history.xml" value="0.33" />
         <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_main.xml" value="0.3625" />
-        <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_selfie.xml" value="0.4" />
+        <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_result_selfie.xml" value="0.40871021775544386" />
+        <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_selfie.xml" value="0.33" />
         <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/activity_take_selfie.xml" value="0.3625" />
         <entry key="..\:/kuliah/Bangkit/fundamental android/OCR_C22-KY03/app/src/main/res/layout/item_row_history.xml" value="0.25" />
       </map>
