+++ app/src/main/AndroidManifest.xml
@@ -14,6 +14,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.OCR_C22KY03">
         <activity
+            android:name=".ResultSelfieActivity"
+            android:exported="false" />
+        <activity
             android:name=".HistoryActivity"
             android:exported="false" />
         <activity
