+++ app/src/main/java/com/bangkit/ocr_c22_ky03/TakeSelfieActivity.kt
@@ -77,10 +77,10 @@ class TakeSelfieActivity : AppCompatActivity() {
                         cameraSelector == CameraSelector.DEFAULT_BACK_CAMERA
                     )
 
-//                    setResult(SelfieActivity.CAMERA_X_RESULT, intent)
+                    setResult(ResultSelfieActivity.CAMERA_X_RESULT, intent)
                     finish()
 
-                    startActivity(Intent(this@TakeSelfieActivity, HistoryActivity::class.java))
+//                    startActivity(Intent(this@TakeSelfieActivity, HistoryActivity::class.java))
                 }
             }
         )
@@ -131,4 +131,6 @@ class TakeSelfieActivity : AppCompatActivity() {
         }
         supportActionBar?.hide()
     }
+
+
 }
