+++ app/src/main/res/layout/activity_selfie.xml
@@ -8,8 +8,8 @@
 
     <TextView
         android:id="@+id/tv_header"
-        android:layout_width="wrap_content"
         style="@style/Text.Header"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
@@ -21,10 +21,10 @@
 
     <TextView
         android:id="@+id/tv_subheader"
+        style="@style/Text.Description"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
-        style="@style/Text.Description"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
         android:text="Make sure there is enough light"
@@ -40,7 +40,7 @@
         android:layout_marginTop="32dp"
         android:layout_marginEnd="16dp"
         android:layout_marginBottom="16dp"
-        app:layout_constraintBottom_toTopOf="@+id/btn_selfie"
+        app:layout_constraintBottom_toTopOf="@+id/btn_next"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tv_subheader"
@@ -48,14 +48,13 @@
         app:srcCompat="@drawable/selfie" />
 
     <Button
-        android:id="@+id/btn_selfie"
-        android:layout_width="230dp"
+        android:id="@+id/btn_next"
         style="@style/Button.Large"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
         android:layout_marginBottom="16dp"
-        android:text="Start Take Selfie"
+        android:text="Next"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
