+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,5 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="primary_blue">#0094DB</color>
+
 </resources>
